+++ packages/native-windows-virtual-gamepad/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./diagnostics";
+export * from "./vhf";
 export * from "./xusb";